fix: address 6 CodeRabbit findings in pr/bug-fixes
- nas/train_cnn_search.py: capture genotype after training step so best_genotype
  matches the architecture that achieved best_valid_acc; initialize to -inf so
  zero-accuracy epochs are correctly recorded
- download_utils.py: guard content-length header against non-numeric values
- config_dict.py: use os.path.isabs() for cross-platform include-path detection
- nas/model.py: derive C_prev from cell.multiplier (actual genotype concat width)
  and validate it matches the network multiplier parameter
- common/utils/utils.py: convert to float before .mean() on stacked tensors
- nas/utils.py: validate gpu_index range before constructing cuda device

+++ tinyml-tinyverse/tinyml_tinyverse/common/utils/utils.py
@@ -716,7 +716,7 @@ class SmoothedValue:
         latest = self.deque[-1]
         if isinstance(latest, torch.Tensor) and latest.ndim == 0:
             d = torch.stack(list(self.deque))
-            return d.mean().item()
+            return d.float().mean().item()
         elif isinstance(latest, numbers.Number):
             d = torch.tensor(list(self.deque), dtype=torch.float32)
             return d.mean().item()
